MessageType now handles registering with a name and payload type.
Fixed bug with memcpy batching.

+++ src/RingBuffer.inl
@@ -11,13 +11,13 @@ namespace dd
 	template <typename T>
 	RingBuffer<T>::RingBuffer()
 	{
-		Allocate( DefaultSize );
+		Allocate(DefaultSize);
 	}
 
 	template <typename T>
-	RingBuffer<T>::RingBuffer( uint size )
+	RingBuffer<T>::RingBuffer(uint size)
 	{
-		Allocate( size );
+		Allocate(size);
 	}
 
 	template <typename T>
@@ -32,32 +32,32 @@ namespace dd
 	}
 
 	template <typename T>
-	void RingBuffer<T>::Push( const T& item )
+	void RingBuffer<T>::Push(const T& item)
 	{
 		uint size = Size();
 		uint capacity = Capacity();
-		if( size == (capacity - 1) )
+		if (size == (capacity - 1))
 		{
 			Grow();
 		}
 
-		new (&m_storage[m_tail]) T( item );
+		new (&m_storage[m_tail]) T(item);
 
 		m_tail = (m_tail + 1) % Capacity();
 		++m_size;
 	}
 
 	template <typename T>
-	void RingBuffer<T>::Push( T&& item )
+	void RingBuffer<T>::Push(T&& item)
 	{
 		uint size = Size();
 		uint capacity = Capacity();
-		if( size == (capacity - 1) )
+		if (size == (capacity - 1))
 		{
 			Grow();
 		}
 
-		new (&m_storage[m_tail]) T( std::move( item ) );
+		new (&m_storage[m_tail]) T(std::move(item));
 
 		m_tail = (m_tail + 1) % Capacity();
 		++m_size;
@@ -66,7 +66,7 @@ namespace dd
 	template <typename T>
 	T RingBuffer<T>::Pop()
 	{
-		DD_ASSERT( m_size != 0 );
+		DD_ASSERT(m_size != 0);
 
 		uint old_head = m_head;
 
@@ -97,18 +97,18 @@ namespace dd
 
 		T* old_data = m_storage.Release();
 
-		Allocate( new_size );
+		Allocate(new_size);
 
-		if( m_tail < m_head )
+		if (m_tail < m_head)
 		{
 			uint front_count = old_size - m_head;
-			CopyRange( old_data + m_head, m_storage.Access(), front_count );
-			CopyRange( old_data, m_storage.Access() + front_count, m_tail );
+			CopyRange(old_data + m_head, m_storage.Access(), front_count);
+			CopyRange(old_data, m_storage.Access() + front_count, m_tail);
 		}
-		else if( m_tail > m_head )
+		else if (m_tail > m_head)
 		{
 			uint count = m_tail - m_head;
-			CopyRange( old_data + m_head, m_storage.Access(), count );
+			CopyRange(old_data + m_head, m_storage.Access(), count);
 		}
 
 		m_head = 0;
@@ -116,11 +116,11 @@ namespace dd
 	}
 
 	template <typename T>
-	void RingBuffer<T>::Allocate( uint size )
+	void RingBuffer<T>::Allocate(uint size)
 	{
-		DD_ASSERT( size > 0 );
+		DD_ASSERT(size > 0);
 
-		m_storage.Set( reinterpret_cast<T*>(new char[size * sizeof( T )]), size );
-		memset( m_storage.Access(), 0xABAD1DEA, size * sizeof( T ) );
+		m_storage.Set(reinterpret_cast<T*>(new char[size * sizeof(T)]), (size_t) size);
+		memset(m_storage.Access(), 0xABAD1DEA, size * sizeof(T));
 	}
 }
